requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
sqlalchemy>=2.0
asyncpg>=0.29
alembic>=1.13
pydantic>=2.6
pydantic-settings>=2.2
python-dotenv>=1.0
passlib[bcrypt]>=1.7
python-jose[cryptography]>=3.3
python-multipart>=0.0.9
email-validator>=2.1
